Specify max reconcile worker to cpu core number. Refactor log system, use high-performance zap library.     Support V(1), V(2), and so on levels     Optimize output format, plus color     Downgrade some trivial output to V(2)     Add -log-level option to command for specifying more verbose output Refactor the logic of updating the status of the security policy.     No need to retry when meeting business logic error     Should retry when it is necessary to wait for the external condition to be established Break up some large files to multiple small files.     The project has tendency to become noodle code, in order to prevent the code from exploding,     refactored some modules, following the idea of doing only one thing at one time, I think we'd be     better to go on refactoring, especially the `builder.go`, it's too long.
