+++ pkg/nsx/services/store_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/vmware/vsphere-automation-sdk-go/runtime/bindings"
 	"github.com/vmware/vsphere-automation-sdk-go/runtime/data"
 	"github.com/vmware/vsphere-automation-sdk-go/services/nsxt/model"
+	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/vmware-tanzu/nsx-operator/pkg/config"
@@ -374,6 +375,77 @@ func Test_transError(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if err := transError(tt.args.err); (err != nil) != tt.wantErr {
 				t.Errorf("transError() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
+	}
+}
+
+func TestListSecurityPolicy(t *testing.T) {
+	groupStore := cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+	policyStore := cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+	ruleStore := cache.NewIndexer(keyFunc, cache.Indexers{util.TagScopeSecurityPolicyCRUID: securityPolicyCRUIDScopeIndexFunc})
+
+	group := model.Group{}
+	scope := "nsx-op/security_policy_cr_uid"
+	uuid := "111111111"
+	id := "1234"
+	group.Id = &id
+	group.UniqueId = &uuid
+
+	group.Tags = []model.Tag{{Scope: &scope, Tag: &id}}
+	groupStore.Add(group)
+
+	id1 := "4567"
+	uuid1 := "111111112"
+	group1 := model.Group{}
+	group1.Id = &id1
+	group1.UniqueId = &uuid1
+	group1.Tags = []model.Tag{{Scope: &scope, Tag: &id1}}
+	groupStore.Add(group1)
+
+	policy := model.SecurityPolicy{}
+	id2 := "1235"
+	policy.Id = &id2
+	policy.UniqueId = &uuid
+	policy.Tags = []model.Tag{{Scope: &scope, Tag: &id2}}
+	policyStore.Add(policy)
+
+	type fields struct {
+		NSXClient           *nsx.Client
+		GroupStore          cache.Indexer
+		SecurityPolicyStore cache.Indexer
+		RuleStore           cache.Indexer
+	}
+	field := fields{NSXClient: nil, GroupStore: groupStore, SecurityPolicyStore: policyStore, RuleStore: ruleStore}
+
+	tests := []struct {
+		name    string
+		fields  fields
+		want    sets.String
+		wantErr bool
+	}{
+		{
+			name:    "test",
+			fields:  field,
+			wantErr: false,
+		},
+	}
+
+	tests[0].want = sets.NewString()
+	tests[0].want.Insert(id)
+	tests[0].want.Insert(id1)
+	tests[0].want.Insert(id2)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			service := &SecurityPolicyService{
+				NSXClient:           tt.fields.NSXClient,
+				GroupStore:          tt.fields.GroupStore,
+				SecurityPolicyStore: tt.fields.SecurityPolicyStore,
+				RuleStore:           tt.fields.RuleStore,
+			}
+			got := service.ListSecurityPolicyID()
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("SecurityPolicyService.ListSecurityPolicyID() = %v, want %v", got, tt.want)
 			}
 		})
 	}
